versão final 05/01 23h20

+++ P2_base/proj_23-24-p2_base/server/eventlist.h
@@ -24,7 +24,7 @@ struct ListNode {
 struct EventList {
   struct ListNode* head;  // Head of the list
   struct ListNode* tail;  // Tail of the list
-  int num_events;
+  size_t num_events;
   pthread_rwlock_t rwl;   // Mutex to protect the list
 };
